DTOs Added
Created Model Package  in  Remote package
Created Repositiories:
-BerryDTO
-BerryDetailDTO
-Berry ResponseDTO
-BerryDetailResponseDTO

Refactored objects w/o DTO added in
-PokemonApiService
-PokeScreen
-BerryDetailViewModel
