+++ app/src/main/java/asdigital/myberryapp/data/remote/model/BerryDTO.kt
@@ -1,0 +1,11 @@
+package asdigital.myberryapp.data.remote.model
+
+data class BerryDTO(
+    val name: String
+)
+
+//data class BerryResponse( val results: List<BerryDTO>)
+
+
+
+
