+++ app/src/main/java/asdigital/myberryapp/data/remote/model/BerryDetailResponseDTO.kt
@@ -1,0 +1,7 @@
+package asdigital.myberryapp.data.remote.model
+
+data class BerryDetailResponseDTO(val berryDetail: List<BerryDetailDTO>)
+
+//data class BerryDetailResponse(val berryDetail: List<BerryDetailDTO>)
+
+
