+++ app/src/main/java/asdigital/myberryapp/data/remote/model/BerryResponseDTO.kt
@@ -1,0 +1,13 @@
+package asdigital.myberryapp.data.remote.model
+
+data class BerryResponseDTO( val results: List<BerryDTO>)
+
+
+fun getBerryDTOList(results: List<BerryDTO>): List<BerryDTO> {
+    return results.map { berry ->
+        BerryDTO(
+            name = berry.name
+        )
+    }
+
+}
