+++ app/src/main/java/asdigital/myberryapp/presentation/BerryDetailViewModel.kt
@@ -3,8 +3,10 @@ package asdigital.myberryapp.presentation
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import asdigital.myberryapp.data.remote.model.BerryDetail
+//import asdigital.myberryapp.data.remote.model.BerryDetail
 import asdigital.myberryapp.data.remote.api.pokeService
+import asdigital.myberryapp.data.remote.model.BerryDetailDTO
+import asdigital.myberryapp.data.remote.model.BerryDetailResponseDTO
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -12,8 +14,8 @@ import kotlinx.coroutines.launch
 
 class BerryDetailViewModel : ViewModel() {
 
-    private val _berryDetailState = MutableStateFlow<BerryDetail?>(null)
-    val berryDetailState: StateFlow<BerryDetail?> = _berryDetailState
+    private val _berryDetailState = MutableStateFlow<BerryDetailDTO?>(null)
+    val berryDetailState: StateFlow<BerryDetailDTO?> = _berryDetailState
 
     private val _errorMessage = MutableStateFlow<String?>(null)
     val errorMessage: StateFlow<String?> = _errorMessage
